code/requirements.txt
anthropic>=0.49.0
voyageai>=0.3.2
chromadb>=0.6.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.32.0
ollama>=0.4.0
